+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## 2.0.4 (unreleased)
 
 
-- Nothing changed yet.
+- Allow to download Knowledge Box logs
 
 
 ## 2.0.3 (2024-02-26)